chore(server): split the parse cache_keys and cached_replay test mods out #3869
module_size_ratchet has been red on main since #3869: cache_keys.rs (458)
and cached_replay.rs (443) crossed the 400-line rule with no allowlist row.
The ratchet's own message says split first and allowlist only with a written
justification, and there is none here: 271 and 295 of those lines are the
inline #[cfg(test)] mod. Both move verbatim to sibling *_tests.rs files,
which the ratchet exempts and which this directory already uses five times
over, including cache_keys_symbolic_tests.rs split out of this very file for
this very reason. cache_keys.rs drops to 185 and cached_replay.rs to 146, so
no row and no digest re-pin is needed.

quality_cache_suffix and cached_geometry_slice widen from private to
pub(super) because their tests now sit one module over. No other non-test
line changes: the pre-mod halves of both files are byte-identical to main
apart from those two keywords, and all 26 tests still run (236 passed).
