+++ apps/server/src/routes/parse/mod.rs
@@ -268,6 +268,12 @@ mod fetch_tests;
 mod cache_keys_symbolic_tests;
 
 #[cfg(test)]
+mod cache_keys_tests;
+
+#[cfg(test)]
+mod cached_replay_tests;
+
+#[cfg(test)]
 mod resolved_tessellation_quality_tests {
     use super::*;
     use crate::error::ApiError;
